Adapt for size_t ByteBuf APIs.

+++ core/Lucy/Index/PostingPool.c
@@ -356,7 +356,7 @@ S_write_terms_and_postings(PostingPool *self, PostingWriter *post_writer,
     ByteBuf *last_term_text
         = BB_new_bytes(post_ivars->blob, post_ivars->content_len);
     char     *last_text_buf  = BB_Get_Buf(last_term_text);
-    uint32_t  last_text_size = BB_Get_Size(last_term_text);
+    uint32_t  last_text_size = (uint32_t)BB_Get_Size(last_term_text);
     SkipStepper_Set_ID_And_Filepos(skip_stepper, 0, 0);
 
     // Initialize sentinel to be used on the last iter, using an empty string
